requirements-dev.txt
aniso8601==10.0.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
flask-restx==1.3.0
gunicorn==23.0.0
identify==2.6.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==25.0
pip-tools==7.4.1
platformdirs==4.3.7
pre_commit==4.2.0
PyMySQL==1.1.1
pyproject_hooks==1.2.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
setuptools==80.1.0
virtualenv==20.30.0
Werkzeug==3.1.3
wheel==0.45.1
